+++ frontend/src/assets/home_page.jpg
