+++ frontend/src/components/SiteFooter.tsx
