+++ frontend/src/styles/designsystem.sass
@@ -47,6 +47,7 @@ $Nunito: "Nunito Sans", sans-serif
   padding: 0
 
 body, button
+  font-size: $text-base
   font-family: $Nunito
 
 .article-content
@@ -97,7 +98,6 @@ body, button
       display: none
       @include responsive($lg)
         display: block
-        font-size: $text-lg
     input, textarea, select, option
       padding: 1rem
       background-color: $primary
