+++ frontend/src/styles/home.sass
@@ -23,12 +23,10 @@
     justify-content: center
     align-items: flex-end
     padding-top: 2rem
-    font-size: $text-3xl
+    font-size: $text-4xl
     font-weight: 600
     font-family: ui-serif, $Cormorant
     letter-spacing: 0.125rem
-    @include responsive($md)
-      font-size: $text-4xl
     @include responsive($lg)
       width: fit
       padding-top: 0rem
@@ -36,12 +34,9 @@
 
   p
     width: 87%
-    font-size: $text-sm
     text-align: center
     line-height: 1.5rem
     padding: 0rem 3rem
-    @include responsive($md)
-      font-size: $text-base
     @include responsive($lg)
       width: 100%
       padding: 0rem 1rem
